refactor: split file-list into drive-data + folder-index modules
